requirements.txt
flask
pre-commit
httpagentparser
pusher
psycopg2-binary
